+++ docs/guide/source/changelog.rst
@@ -42,6 +42,21 @@ Unreleased [3.0.0] (MuJoCo 3.5.0)
 
     - :docs-rs:`~mujoco_rs::wrappers::mj_data::<struct>MjData::<method>forward_kinematics`.
 
+- New methods:
+
+  - |mj_model|:
+
+    - :docs-rs:`~mujoco_rs::wrappers::mj_model::<struct>MjModel::<method>extract_state`
+      :docs-rs:`~mujoco_rs::wrappers::mj_model::<struct>MjModel::<method>extract_state_into`.
+
+- Changed methods:
+
+  - |mj_spec|:
+
+    - :docs-rs:`~mujoco_rs::wrappers::mj_editing::<struct>MjSpec::<method>from_buffer`
+      now calls :docs-rs:`~~mujoco_rs::mujoco_c::<struct>mj_loadModelBuffer` directly.
+      A virtual file-system was previously used to support this.
+
 - Other changes:
 
   - Updated enum type aliases.